clean up, secure and document accounts API and tokens

+++ server/routes/accounts.js
@@ -1,20 +1,28 @@
+// TODO hash passwords
 var db = require('../data/db.js');
 
 // get accounts
 module.exports.get = function (req, res) {
-  if (!req.params.username) {
-    // TODO only admin
+
+  // check whether requesting user has permissions
+  console.log(req.decodedToken.username + " is admin? " + req.decodedToken.isAdmin);
+  if (!req.decodedToken.isAdmin) {
+    res.status(403).json({"message": "not authorized"});
+  }
+  else if (!req.params.username) {
     // get all
     db.table('accounts')
-      // todo change from all to all for user
       .run()
       .then(function (result) {
         if (result.length == 0) {
           res.status(404).json({error: "no accounts found"});
         }
         else {
-          // TODO just return names
-          res.json(result);
+          res.json({
+            "id": result.id,
+            "username": result.username,
+            "isAdmin": result.isAdmin
+          });
         }
       })
       .error(function (err) {
@@ -35,7 +43,11 @@ module.exports.get = function (req, res) {
           res.status(404).json({error: "account not found"});
         }
         else {
-          res.json(result);
+          res.json({
+            "id": result.id,
+            "username": result.username,
+            "isAdmin": result.isAdmin
+          });
         }
       })
       .error(function (err) {
@@ -46,21 +58,26 @@ module.exports.get = function (req, res) {
 };
 
 
-// TODO only admin
 // create account
 module.exports.post = function (req, res) {
-	var username = req.body.username;
+  var username = req.body.username;
   var password = req.body.password;
   var isAdmin = req.body.isAdmin;
 
-  if (!username) {
-		res.json({error:"must provide username"});
-	}
-	if (!password) {
-		res.json({error:"must provide password"});
-	}
+  // check whether requesting user has permissions
+  console.log(req.decodedToken.username + " is admin? " + req.decodedToken.isAdmin);
+  if (!req.decodedToken.isAdmin) {
+    res.status(403).json({"message": "not authorized"});
+  }
 
-	// check for existing user with same name, otherwise add
+  if (!username) {
+    res.json({error: "must provide username"});
+  }
+  if (!password) {
+    res.json({error: "must provide password"});
+  }
+
+  // check for existing user with same name, otherwise add
   db.table('accounts')
     .filter({
       "username": username
@@ -80,27 +97,31 @@ module.exports.post = function (req, res) {
     db.table('accounts')
       .insert(
         {
-          "username" : username,
-          "password" : password,
-          "isAdmin" : isAdmin
+          "username": username,
+          "password": password,
+          "isAdmin": isAdmin
         })
       .run()
-      .then(function(response) {
-        res.json(response);
+      .then(function (result) {
+        res.json({
+          "username": username
+        });
       })
-      .error(function(err) {
+      .error(function (err) {
         res.json({error: err});
       })
   }
 
 };
 
-// TODO only update own user
 // update a user by username
 module.exports.put = function (req, res) {
   if (!req.params.username) {
     res.status(400).json({"error": "must provide username"});
-    // TODO should check password?
+  }
+  // can only be updated by admin or self
+  else if (req.decodedToken.username != req.params.username || !req.decodedToken.isAdmin) {
+    res.status(403).json({"message": "not authorized"});
   }
   else {
     db.table('accounts')
@@ -110,10 +131,11 @@ module.exports.put = function (req, res) {
       .update({
         "username": req.body.username,
         "password": req.body.password,
+        "isAdmin": req.body.isAdmin
       })
       .then(function (result) {
         if (result.replaced == 1 || result.unchanged == 1) {
-          res.json({"username": req.params.username});
+          res.json({"username": req.body.username});
         }
         else {
           res.status(404).json({error: "account not found"});
@@ -126,30 +148,36 @@ module.exports.put = function (req, res) {
   }
 };
 
-// TODO only admin
 // delete a user by username
 module.exports.delete = function (req, res) {
   if (!req.params.username) {
     res.status(400).json({"error": "must provide username"});
   }
   else {
-    db.table('accounts')
-      .filter({
-        "username": req.params.username
-      })
-      .delete()
-      .then(function (result) {
-        if (result.deleted != 1) {
-          res.status(404).json({error: "username not found"});
-        }
-        else {
-          res.json({"deleted": req.params.username});
-        }
-      })
-      .error(function (err) {
-        console.log(err);
-        res.status(500).json({"error": "an error occurred deleting the record"});
-      })
+
+    console.log(req.decodedToken.username + " is admin? " + req.decodedToken.isAdmin);
+    if (!req.decodedToken.isAdmin) {
+      res.status(403).json({"message": "not authorized"});
+    }
+    else {
+      db.table('accounts')
+        .filter({
+          "username": req.params.username
+        })
+        .delete()
+        .then(function (result) {
+          if (result.deleted != 1) {
+            res.status(404).json({error: "username not found"});
+          }
+          else {
+            res.json({"deleted": req.params.username});
+          }
+        })
+        .error(function (err) {
+          console.log(err);
+          res.status(500).json({"error": "an error occurred deleting the record"});
+        })
+    }
   }
 };
